myproject/requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
pytest==7.4.2